[node-loader] Fix single and multi ID loader to use node type
Previously, calling getByID on a Node class would load nodes of any class that had that ID. This is incorrect behavior. Calling getByID on a subclass of GPNode should only load nodes of that type. To load nodes of any type, use GPNode::getByID.

Possible breaking changes:
If you relied on this functionality, make sure to replace the object classed on ::getByID and ::multiGetByID to GPNode. (e.g. User::getByID becomes GPNode::getByID)